+++ src/core/base64.h
@@ -8,7 +8,7 @@
 
 #include <string>
 
-std::string base64_encode(unsigned char const* , unsigned int len);
+std::string base64_encode(unsigned char const* , size_t len);
 std::string base64_decode(std::string const& s);
 
 #endif /* BASE64_H */
